[C] Sweep for flow and congestion control suppliers for channel length type.

+++ aeron-driver/src/main/c/aeron_flow_control.c
@@ -98,9 +98,7 @@ int64_t aeron_max_flow_control_strategy_on_sm(
     return snd_lmt > window_edge ? snd_lmt : window_edge;
 }
 
-bool aeron_max_flow_control_strategy_should_linger(
-    void *state,
-    int64_t now_ns)
+bool aeron_max_flow_control_strategy_should_linger(void *state, int64_t now_ns)
 {
     aeron_max_flow_control_strategy_state_t *strategy_state = (aeron_max_flow_control_strategy_state_t *)state;
 
@@ -119,7 +117,7 @@ int aeron_max_flow_control_strategy_fini(aeron_flow_control_strategy_t *strategy
 
 int aeron_max_multicast_flow_control_strategy_supplier(
     aeron_flow_control_strategy_t **strategy,
-    int32_t channel_length,
+    size_t channel_length,
     const char *channel,
     int32_t stream_id,
     int64_t registration_id,
@@ -151,7 +149,7 @@ int aeron_max_multicast_flow_control_strategy_supplier(
 
 int aeron_unicast_flow_control_strategy_supplier(
     aeron_flow_control_strategy_t **strategy,
-    int32_t channel_length,
+    size_t channel_length,
     const char *channel,
     int32_t stream_id,
     int64_t registration_id,
