+++ aeron-driver/src/main/c/aeronmd.h
@@ -244,7 +244,7 @@ typedef struct aeron_flow_control_strategy_stct aeron_flow_control_strategy_t;
 
 typedef int (*aeron_flow_control_strategy_supplier_func_t)(
     aeron_flow_control_strategy_t **strategy,
-    int32_t channel_length,
+    size_t channel_length,
     const char *channel,
     int32_t stream_id,
     int64_t registration_id,
